Make sense of the two `Description` components
Rename one to make their difference clear, add docs and move them into
the same file.
