+++ frontend/src/routes/Search.tsx
@@ -10,7 +10,7 @@ import { loadQuery } from "../relay";
 import { Link } from "../router";
 import { Creators, isPastLiveEvent, Thumbnail } from "../ui/Video";
 import { unreachable } from "../util/err";
-import { Description } from "../ui/metadata";
+import { SmallDescription } from "../ui/metadata";
 import { Card } from "../ui/Card";
 import { PageTitle } from "../layout/header/ui";
 import { BreadcrumbsContainer, BreadcrumbSeparator } from "../ui/Breadcrumbs";
@@ -219,7 +219,7 @@ const SearchEvent: React.FC<SearchEventProps> = ({
                     WebkitLineClamp: 2,
                 }}>{title}</h3>
                 <Creators creators={creators} />
-                <Description text={description} lines={3} />
+                <SmallDescription text={description} lines={3} />
                 {/* TODO: link to series */}
                 {seriesTitle && <div css={{ fontSize: 14, marginTop: 4 }}>
                     {t("video.part-of-series") + ": " + seriesTitle}
